--- demo/src/main/webapp/WEB-INF/web.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="ISO-8859-1"?>
-<web-app xmlns="http://java.sun.com/xml/ns/j2ee"
-	 xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-	 xsi:schemaLocation="http://java.sun.com/xml/ns/j2ee http://java.sun.com/xml/ns/j2ee/web-app_2_4.xsd"
-	 version="2.4">
-
-	<display-name>demo</display-name>
-
-	 <!--  
-	      There are three means to configure Wickets configuration mode and they are
-	      tested in the order given. 
-	      1) A system property: -Dwicket.configuration
-	      2) servlet specific <init-param>
-	      3) context specific <context-param>
-	      The value might be either "development" (reloading when templates change)
-	      or "deployment". If no configuration is found, "development" is the default.
-	-->
-
-	<filter>
-		<filter-name>wicket.demo</filter-name>
- 		<filter-class>org.apache.wicket.protocol.http.WicketFilter</filter-class>
-		<init-param>
-			<param-name>applicationClassName</param-name>
-			<param-value>com.jolira.stateless.WicketApplication</param-value>
- 		</init-param>
- 	</filter>
-
- <filter-mapping>
-  <filter-name>wicket.demo</filter-name>
-	<url-pattern>/*</url-pattern>
- </filter-mapping>
-
-
-</web-app>
