+++ pom.xml
@@ -1,113 +1,25 @@
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <groupId>org.wicketstuff</groupId>
-  <artifactId>wicketstuff-stateless-parent</artifactId>
-  <packaging>pom</packaging>
-  <version>6.6.0-SNAPSHOT</version>
-  <name>WicketStuff Stateless Parent</name>
-  <description>Components for Apache Wicket useful for Stateless Applications</description>
-  <properties>
-    <wicket.version>6.6.0</wicket.version>
-    <slf4j.version>1.7.2</slf4j.version>
-    <jetty.version>7.6.9.v20130131</jetty.version>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-  </properties>
-  <licenses>
-    <license>
-      <name>The Apache Software License, Version 2.0</name>
-      <url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
-      <distribution>repo</distribution>
-    </license>
-  </licenses>
-    <modules>
-        <module>stateless</module>
-        <module>scala-demo</module>
-        <module>demo</module>
-    </modules>
-  <dependencyManagement>
-    <dependencies>
-      <dependency>
-        <groupId>org.apache.wicket</groupId>
-        <artifactId>wicket-core</artifactId>
-        <version>${wicket.version}</version>
-      </dependency>
-      <dependency>
-        <groupId>junit</groupId>
-        <artifactId>junit</artifactId>
-        <version>4.11</version>
-        <scope>test</scope>
-      </dependency>
+	<modelVersion>4.0.0</modelVersion>
 
-      <!-- LOGGING DEPENDENCIES -->
-      <dependency>
-        <groupId>org.slf4j</groupId>
-        <artifactId>slf4j-simple</artifactId>
-        <version>${slf4j.version}</version>
-      </dependency>
+	<parent>
+		<groupId>org.wicketstuff</groupId>
+		<artifactId>jdk-1.6-parent</artifactId>
+		<version>6.0-SNAPSHOT</version>
+	</parent>
 
-      <!--  JETTY DEPENDENCIES FOR TESTING  -->
-      <dependency>
-        <groupId>org.eclipse.jetty</groupId>
-        <artifactId>jetty</artifactId>
-        <version>${jetty.version}</version>
-        <scope>provided</scope>
-      </dependency>
-      <dependency>
-        <groupId>org.eclipse.jetty</groupId>
-        <artifactId>jetty-util</artifactId>
-        <version>${jetty.version}</version>
-        <scope>provided</scope>
-      </dependency>
-      <dependency>
-        <groupId>org.eclipse.jetty</groupId>
-        <artifactId>jetty-management</artifactId>
-        <version>${jetty.version}</version>
-        <scope>provided</scope>
-      </dependency>
-    </dependencies>
-  </dependencyManagement>
-  <build>
-    <resources>
-      <resource>
-        <filtering>false</filtering>
-        <directory>src/main/resources</directory>
-      </resource>
-      <resource>
-        <filtering>false</filtering>
-        <directory>src/main/java</directory>
-        <includes>
-          <include>**</include>
-        </includes>
-        <excludes>
-          <exclude>**/*.java</exclude>
-        </excludes>
-      </resource>
-    </resources>
-    <testResources>
-      <testResource>
-        <filtering>false</filtering>
-        <directory>src/test/java</directory>
-        <includes>
-          <include>**</include>
-        </includes>
-        <excludes>
-          <exclude>**/*.java</exclude>
-        </excludes>
-      </testResource>
-    </testResources>
-    <plugins>
-      <plugin>
-        <inherited>true</inherited>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-compiler-plugin</artifactId>
-        <version>2.5.1</version>
-        <configuration>
-          <source>1.6</source>
-          <target>1.6</target>
-          <optimize>true</optimize>
-          <debug>true</debug>
-        </configuration>
-      </plugin>
-    </plugins>
-  </build>
+	<artifactId>wicketstuff-stateless-parent</artifactId>
+	<packaging>pom</packaging>
+	<name>WicketStuff Stateless Parent</name>
+	<description>Components for Apache Wicket useful for Stateless Applications</description>
+	<licenses>
+		<license>
+			<name>The Apache Software License, Version 2.0</name>
+			<url>http://www.apache.org/licenses/LICENSE-2.0.txt</url>
+			<distribution>repo</distribution>
+		</license>
+	</licenses>
+	<modules>
+		<module>stateless</module>
+		<module>stateless-examples</module>
+	</modules>
 </project>
