--- scala-demo/README
@@ -1,23 +1,0 @@
-A demo project for the stateless components and behaviors.
-
-Written in Scala just because I'm learning the language.
-
-This project can be built with either Maven or Simple Build Tool (SBT).
-To run with sbt, install SBT, run sbt in this directory and type 'update' then jetty-run at the prompt.
-
-Or to run the tests:
-sbt update && sbt test
-mvn test
-
-See the demo project (sibling of this) for more examples of usage of the library.
-
-Usage of stateless Ajax components and behaviors:
-
- - the components which have to be repainted should have stable/hardcoded
- markup ids
- This is needed because every request creates a new instance of the page and re-creates
- all components in it, thus the auto-assigned markup id by Wicket can be different
- than the one assigned at the previous page/component render.
- Example:
-	Label label = new Label("l1", "Model");
-	label.setMarkupId("somethingStable");
