--- scala-demo/project/build.properties
@@ -1,9 +1,0 @@
-#
-#Mon Feb 21 11:11:37 EET 2011
-project.name=myproject
-project.organization=com.mycompany
-scala.version=2.8.1
-project.version=1.0-SNAPSHOT
-sbt.version=0.7.5.RC0
-build.scala.versions=2.8.1
-project.initialize=false
