--- scala-demo/src/main/scala/org/wicketstuff/stateless/AjaxEventPage.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Ajax event demo</title>
-  </head>
-
-  <body>
-
-    <span wicket:id="l1">label1</span>
-    <br/>
-    <span wicket:id="l2">label1</span>
-
-  </body>
-
-</html>
