--- scala-demo/src/main/scala/org/wicketstuff/stateless/HomePage.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-    <style type="text/css">
-        INPUT, BODY { font-size: 24px; font-family: Helvetica; }
-        BODY { margin-left: 30%; margin-right: 30%; margin-top: 10%; }
-        .header { background-color: #E05717; color: white; padding: 10px; }
-        FORM { background-color: #7EACCB; padding: 30px; }
-        P { margin: 0 }
-    </style>
-</head>
-
-<body>
-  <span wicket:id="message">default</span>
-</body>
-
-</html>
