--- scala-demo/src/main/scala/org/wicketstuff/stateless/LinkPage.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <title>Stateless Link demo</title>
-  </head>
-
-  <body>
-
-    <a wicket:id="link">click me</a>
-
-  </body>
-</html>
