--- scala-demo/src/main/scala/org/wicketstuff/stateless/LinkPage.scala
@@ -1,22 +1,0 @@
-package org.wicketstuff.stateless
-
-import org.apache.wicket.request.mapper.parameter.PageParameters
-import org.apache.wicket.markup.html.WebPage
-import com.google.code.joliratools.StatelessLink
-import org.apache.wicket.model.Model
-
-/**
- * A page that shows usage of StatelessLink
- */
-class LinkPage(parameters: PageParameters)
-  extends WebPage(parameters) {
-
-  val link = new StatelessLink("link", Model.of("Empty"), parameters) {
-
-    override def onClick {
-      println("clicked!")
-    }
-  };
-  add(link);
-
-}
