--- scala-demo/src/main/webapp/WEB-INF/web.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE web-app
-      PUBLIC "-//Sun Microsystems, Inc.//DTD Web Application 2.3//EN"
-      "http://java.sun.com/dtd/web-app_2_3.dtd">
-
-<web-app>
-    <display-name>Stateless Demo</display-name>
-    <filter>
-        <filter-name>Stateless</filter-name>
-        <filter-class>org.apache.wicket.protocol.http.WicketFilter</filter-class>
-        <init-param>
-          <param-name>applicationClassName</param-name>
-          <param-value>org.wicketstuff.stateless.StatelessApplication</param-value>
-        </init-param>
-    </filter>
-    <filter-mapping>
-        <filter-name>Stateless</filter-name>
-        <url-pattern>/*</url-pattern>
-    </filter-mapping>
-</web-app>
