+++ stateless-examples/src/main/resources/log4j.properties
