+++ stateless-examples/src/test/java/org/wickststuff/stateless/demo/Start.java
@@ -1,0 +1,58 @@
+package org.wickststuff.stateless.demo;
+
+import org.eclipse.jetty.server.Connector;
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.server.bio.SocketConnector;
+import org.eclipse.jetty.webapp.WebAppContext;
+
+/**
+ * @author jfk
+ * @date Jan 27, 2011 11:49:02 AM
+ * @since 1.0
+ * 
+ */
+public class Start {
+
+    /**
+     * @param args
+     * @throws Exception
+     */
+    public static void main(final String[] args) throws Exception {
+        final Server server = new Server();
+        final SocketConnector connector = new SocketConnector();
+
+        // Set some timeout options to make debugging easier.
+        connector.setMaxIdleTime(1000 * 60 * 60);
+        connector.setSoLingerTime(-1);
+        connector.setPort(8080);
+        server.setConnectors(new Connector[] { connector });
+
+        final WebAppContext bb = new WebAppContext();
+        bb.setServer(server);
+        bb.setContextPath("/");
+        bb.setWar("src/main/webapp");
+
+        // START JMX SERVER
+        // MBeanServer mBeanServer = ManagementFactory.getPlatformMBeanServer();
+        // MBeanContainer mBeanContainer = new MBeanContainer(mBeanServer);
+        // server.getContainer().addEventListener(mBeanContainer);
+        // mBeanContainer.start();
+
+        server.setHandler(bb);
+
+        try {
+            System.out.println(">>> STARTING EMBEDDED JETTY SERVER, PRESS ANY KEY TO STOP");
+            server.start();
+            System.in.read();
+            System.out.println(">>> STOPPING EMBEDDED JETTY SERVER");
+            // while (System.in.available() == 0) {
+            // Thread.sleep(5000);
+            // }
+            server.stop();
+            server.join();
+        } catch (final Exception e) {
+            e.printStackTrace();
+            System.exit(100);
+        }
+    }
+}
