+++ stateless-examples/src/test/java/org/wickststuff/stateless/demo/TestHomePage.java
@@ -1,0 +1,26 @@
+package org.wickststuff.stateless.demo;
+
+import junit.framework.TestCase;
+import org.apache.wicket.util.tester.WicketTester;
+import org.wicketstuff.stateless.demo.HomePage;
+import org.wicketstuff.stateless.demo.WicketApplication;
+
+/**
+ * Simple test using the WicketTester
+ */
+public class TestHomePage extends TestCase {
+    private WicketTester tester;
+
+    @Override
+    public void setUp() {
+        tester = new WicketTester(new WicketApplication());
+    }
+
+    public void testRenderMyPage() {
+        // start and render the test page
+        tester.startPage(HomePage.class);
+
+        // assert rendered page class
+        tester.assertRenderedPage(HomePage.class);
+    }
+}
